+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 version 0.4
 - added get_references() function
 - removed dependancy on DateTime perl module in compile_query()
-- iri_crunch() function added
+- iri_crunch() function added to convert predicates into readable form
+- added 'schema' [http://schema.org/]  namespace
+- added openarchives.org namespaces 'oai','ore','oai_dc' 
+- added other wanted namespaces: 'xsi', 'dcat'
 
 version 0.3
 - get_properties() function improvements